[public-api] Move oidc handler package from iam
